# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fountainrank-backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/fountains/{fountain_id}/place": {
      "get": {
        "tags": [
          "fountains"
        ],
        "summary": "Fountain Place",
        "description": "One fountain's PUBLIC place membership + the \u00a77 indexing verdict (spec \u00a75/\u00a77).\n\nPublic + unauthenticated + cacheable, computed only from non-hidden columns (never the\nviewer/admin detail path), so auth/admin state can never influence indexability or SEO copy.\nResolves the fountain's precomputed country/city place from the membership columns (never a\nlive ST_Covers). 404s a hidden or unknown fountain (matching the detail endpoint), so the web\npage can ``noindex`` + ``notFound()``. ``indexable`` is the single \u00a77 predicate, evaluated in\nthe same query that loads the row so it can never drift from the sitemap enumeration.",
        "operationId": "fountain_place_api_v1_fountains__fountain_id__place_get",
        "parameters": [
          {
            "name": "fountain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Fountain Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FountainPlaceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/places": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "List Places",
        "description": "List crawlable places, most-populous first.\n\nCountries when ``country`` is omitted; the cities of ``country`` otherwise. Only **canonical**\nplaces (one per ``(country_code, slug)``) with ``fountain_count >= seo_place_min_fountains``\nare returned \u2014 the thin-content gate (spec \u00a77). Hidden fountains are already excluded from\n``fountain_count`` (denormalized at membership refresh), so no per-row hidden filter is needed\nhere. Bounded by ``limit``/``offset`` and served with a public ``Cache-Control`` (both in the\ncontract, not just tests).",
        "operationId": "list_places_api_v1_places_get",
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-3166-1 alpha-2 country code. Omit to list countries; provide it to list that country's cities (its canonical child places).",
              "title": "Country"
            },
            "description": "ISO-3166-1 alpha-2 country code. Omit to list countries; provide it to list that country's cities (its canonical child places)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaceOut"
                  },
                  "title": "Response List Places Api V1 Places Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/places/{country}/{city}/fountains": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "City Fountains",
        "description": "The canonical city for /[country]/[city] plus its ranked, paginated fountains (spec \u00a74.3/\u00a75).\n\nResolves the canonical place owning ``(country_code, slug)`` (both matched lowercased \u2014 slugs\nare stored lowercased), then returns its NON-HIDDEN fountains best-rated first (Bayesian\n``ranking_score`` DESC, unrated last). Reads the precomputed ``city_place_id`` membership, not a\nlive ST_Covers (spec \u00a75). Public + cacheable. 404 when no canonical city matches, so the page\ncan ``notFound()``.",
        "operationId": "city_fountains_api_v1_places__country___city__fountains_get",
        "parameters": [
          {
            "name": "country",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Country"
            }
          },
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityFountainsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CityFountainsOut": {
        "properties": {
          "place": {
            "$ref": "#/components/schemas/PlaceOut"
          },
          "fountains": {
            "items": {
              "$ref": "#/components/schemas/FountainPin"
            },
            "type": "array",
            "title": "Fountains"
          },
          "indexable": {
            "type": "boolean",
            "title": "Indexable"
          }
        },
        "type": "object",
        "required": [
          "place",
          "fountains",
          "indexable"
        ],
        "title": "CityFountainsOut",
        "description": "A city place plus its ranked, paginated fountains (#127 Slice 3, spec \u00a74.3/\u00a75).\n\n``place`` is the canonical city that owns the ``/[country]/[city]`` URL; ``fountains`` are its\nnon-hidden fountains, best-rated first. ``place.fountain_count`` is the full non-hidden total\n(the list is capped by ``limit``), so the page can show \"top N of M\" without a separate count.\n``indexable`` is the spec \u00a77 thin-content predicate computed server-side (``fountain_count >=\nseo_place_min_fountains``) \u2014 the single source of truth for K, so the web sets ``noindex`` from\nit rather than knowing the threshold."
      },
...
      "Coordinates": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "Coordinates"
      },
...
      "FountainPin": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "location": {
            "$ref": "#/components/schemas/Coordinates"
          },
          "is_working": {
            "type": "boolean",
            "title": "Is Working"
          },
          "average_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Rating"
          },
          "rating_count": {
            "type": "integer",
            "title": "Rating Count"
          },
          "ranking_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ranking Score"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "current_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Status"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "location",
          "is_working",
          "average_rating",
          "rating_count"
        ],
        "title": "FountainPin"
      },
      "FountainPlaceOut": {
        "properties": {
          "fountain_id": {
            "type": "string",
            "format": "uuid",
            "title": "Fountain Id"
          },
          "city": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "country": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "indexable": {
            "type": "boolean",
            "title": "Indexable"
          }
        },
        "type": "object",
        "required": [
          "fountain_id",
          "city",
          "country",
          "indexable"
        ],
        "title": "FountainPlaceOut",
        "description": "One fountain's public place membership + indexability verdict (#127 Slice 5, spec \u00a75/\u00a77).\n\nComputed from PUBLIC, non-hidden data only (never the viewer/admin path), so auth/admin state\ncan never influence indexability or SEO copy. ``city`` is the fountain's most-specific covering\ncity place \u2014 it shares its ``(country_code, slug)`` with the canonical city that owns the public\n``/[country]/[city]`` URL \u2014 and ``country`` is its country place; either is ``None`` when\nunmatched. Read from the precomputed membership columns (never a live ST_Covers, spec \u00a75).\n``indexable`` is the single server-side \u00a77 predicate, so the web sets ``noindex`` from it\nwithout re-deriving the rule."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaceOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "country_code": {
            "type": "string",
            "title": "Country Code"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "subtype": {
            "type": "string",
            "title": "Subtype"
          },
          "fountain_count": {
            "type": "integer",
            "title": "Fountain Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "country_code",
          "slug",
          "name",
          "subtype",
          "fountain_count"
        ],
        "title": "PlaceOut",
        "description": "A crawlable place (country or city) for the public SEO endpoints (#127, spec \u00a75).\n\nSerialized from ``PlaceBoundary`` (precomputed membership; never a live ST_Covers). The\npublic URL segment is ``country_code`` for a country (ISO-3166-1 alpha-2, lowercased) and\n``slug`` for a city \u2014 both are carried so the client can build either route. ``fountain_count``\nis the denormalized NON-HIDDEN count that drives the \"N fountains\" copy and the >= K gate."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}